requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
python-dotenv==1.0.1
WTForms==3.1.2
email-validator==2.1.0.post1
requests==2.31.0
openai==1.12.0
Pillow==10.3.0
gTTS==2.5.1
SpeechRecognition==3.10.1
python-twitter==3.5
facebook-sdk==3.1.0
linkedin-api==2.0.0a5
instabot==0.117.0
medium-api==0.1.0
pandas==2.2.1 